+++ cmd/metal-api/internal/service/partition-service.go
@@ -171,7 +171,7 @@ func (r partitionResource) createPartition(request *restful.Request, response *r
 	if requestPayload.MgmtServiceAddress != nil {
 		mgmtServiceAddress = *requestPayload.MgmtServiceAddress
 	}
-	prefixLength := 22
+	prefixLength := uint8(22)
 	if requestPayload.PrivateNetworkPrefixLength != nil {
 		prefixLength = *requestPayload.PrivateNetworkPrefixLength
 		if prefixLength < 16 || prefixLength > 30 {
